code/requirements.txt
sentence-transformers
sagemaker-containers
numpy>=1.17
